feat: add `EventQueue` class util to guarantee the order of user-storage updates (#5937)
## Explanation

This PR add a new `EventQueue` shared class util to the
`@metamask/profile-sync-controller` package.
We are using this to ensure user-storage updates are executed and
fulfilled in the actual order they were called.
For now, this is used for account syncing when reacting to
`AccountsController` `accountRenamed` and `accountAdded` events.

This addition has been made in reaction to a recent change to extension
code that made it so that when a user adds a new account, without
changing the default name, both `accountAdded` and `accountRenamed` were
fired.
This induced race conditions where we weren't in control of which
user-storage request would land first.
With this change, we raise back our confidence to 100%.

- ✅ Extension test drive PR:
https://github.com/MetaMask/metamask-extension/pull/33552

## References

* Fixes: https://consensyssoftware.atlassian.net/browse/IDENTITY-124

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes