network-controller: Represent dependency on ErrorReportingService more accurately
A previous commit that updated NetworkController to auto-correct
`selectedNetworkClientId` also made it so that the controller now
expects the client to initialize ErrorReportingService and its messenger
ahead of time (and for `ErrorReportingService:captureException` to be
added to NetworkControllerMessenger's allowlist). This fact was never
enforced properly, and it was never documented in the changelog.

This commit addresses this issue by moving
`@metamask/error-reporting-service` to `peerDependencies` and bringing
the changelog up to date.