fix: prevent `TokenBalancesController` updating account balance to 0 while multicall contract failed (#5975)
## Explanation

This PR fix:
TokenBalancesController updating account balance to 0 while multicall
contract failed

in `AssetsControllers/TokenBalancesController` ->
`updateBalancesByChainId`
it use multicall contract to continually poll the balance and update the
balance for the account,
however even if the contract request failed, it will continue to update

the root cause is because the multicall return an array of
MulticallResult object even it failed, but the value of the balance may
become undefined

hence, the `updateBalancesByChainId` will continue using this value to
process the logic, and result update the balance to 0


Attached the extension screen that impacted by above logic


![image](https://github.com/user-attachments/assets/8c7e1f63-0bd5-49a6-bc6a-92f7db9a014f)


<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes