Removed `recoveryRatelimitCache` from the Seedless controller state (#5976)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
This PR removes the `recoveryRatelimitCache` data from the controller
state.
Previously, the state was needed to address the state sync between
devices because of the inaccurate rate-limit blocking time from the
server. Recently, the fixes for this were addressed in the Server as
well as in the SDK, hence the state data are redundant and will not work
anymore.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: Chaitanya Potti <chaitanya.potti@gmail.com>
Co-authored-by: himanshuchawla009 <himanshuchawla2014@gmail.com>
Co-authored-by: matthiasgeihs <62935430+matthiasgeihs@users.noreply.github.com>
Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>
Co-authored-by: Nguyen Anh Tu <tunguyenanh.tech@gmail.com>
Co-authored-by: ieow <cwlee0@gmail.com>
Co-authored-by: Nguyen Anh Tu <tunguyen.bhtech@gmail.com>
Co-authored-by: Tuna <tunguyenanh@tunatech.org>