Reduce default circuit break duration to 2 minutes
The `createServicePolicy` utility sets the default "circuit break
duration" to 30 minutes. When used to hit an API, this signifies the
amount of time requests to that API that will be paused if it
consistently responds with enough 500s. Since the `createServicePolicy`
utility is used to make network requests, however, it also signifies the
amount of time that users will be "locked out" from interacting with a
blockchain until requests respond successfully.

Pausing for 30 minutes makes sense if an RPC endpoint is truly down, but
it creates a major inconvenience if the endpoint mostly works and only
certain kinds of requests fail. To address this, this commit reduces the
default circuit break duration to 2 minutes to allow for faster
recovery.