Reduce default circuit break duration to 2 minutes (#6015)
Currently, the "circuit break duration" — the duration for which
requests to an RPC endpoint are paused when the endpoint is perceived to
be unavailable — is 30 minutes. This is a good value when a true outage
exists, but if the endpoint is merely degraded or is returning 500s for
specific requests then we don't want to lock users out of their wallet
for a long time. To accommodate these cases, this commit reduces the
circuit break duration to 2 minutes so that recovery is faster.