fix: update gasLimit to be passed as Hex string or undefined (#6038)
## Explanation



* What is the current state of things and why does it need to change?
- Currently we incorrectly pass gasLimit as a stringified number when
the addTransaction method expects a hex string
- Currently the Transactions controller allows gasLimit to be undefined,
passing the gas estimation task to the Transaction controller, we want
to utilize this ability in the future

* What is the solution your changes offer and how does it work?
- For any contract method that uses a gasLimit, allow
gasOptions.gasLimit to be set to 'none' and translate that to setting
gasLimit undefined in addTransaction params
- In the case that there is a gasLimit convert it to Hex string

* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
- Yes. The stake-sdk contract methods take gasLimit params such that gas
can be estimated if needed. The params allow gas to be set to none, but
that equates to 0 gas. We do not need to typically send 0 gas in our
cases. We can instead say that if one passes gasLimit none, in the earn
controller it means that they want to set addTransaction gasLimit to
undefined which will pass the task to Transaction Controller

- <!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

Fixes https://consensyssoftware.atlassian.net/browse/TAT-1064

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes