Rename team-solana label to team-new-networks (#5951)
The `team-solana` issue label in the extension and mobile repos has been
renamed to `team-new-networks`. The `create-update-issues` workflow uses
this to know which teams to assign to tickets representing controller
upgrades. Updating this label in `teams.json` ensures that this workflow
won't break when a new major version of
`@metamask/multichain-transactions-controller` is released.