fix: Auth Pubkey persistance in the Seedless controller state (#6055)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
In the SeedlessOnboardingController, we stored the `AuthPubKey` in the
controller state, to be used in the password-change sync flow (i.e.
password change can be detected by comparing the latest AuthPubKey from
the server against the local AuthPubKey stored in the controller state).
Currently we have a method to persist the AuthPubKey,
`#persistAuthPubKey`, which updates the AuthPubKey state whenever (new)
vault is created or updated with new AuthData from the server.
The issue that `#persistAuthPubKey` isn't being called on every vault
update, even though the data should be updated whenever `Vault` is
updated.

This PR addresses the incorrect implementation for persisting
`AuthPubKey` in the controller state as well as updated the tests with
correct mocks and order of mocks.
 
**Previously, we weren't able to detect this issue due to the incorrect
mocks in the test cases.**

## References
Thanks to this comment:
https://github.com/MetaMask/core/pull/6047#discussion_r2176529855
<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes