feat: keep correct order of secret data backups (SRP and Private Key) (#6047)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
This pr mainly addresses the following issues regarding to the import
order of Different SecretData backup ~
The current implementation of `fetchAllSecretData` method returns the
both Srp (Mnemonics) and Private key backup in object mapping data
structure, i.e.
```ts
{
  Mnemonics: [<Mnemonic_1>, <Mnemonic_2>, <Mnemonic_3>],
  PrivateKey: [<PrivateKey_1>, <PrivateKey_2>],
}
```
This data structure preserves the order of each SecretType, `Mnemonics`
and `PrivateKey` (in an array) but that doesn't solves the order of
backup created for all SecretTypes. For instance,
user created backups in order of `Mnemonic_1`->`Mnemonic_2` ->
`PrivateKey_1` -> `PrivateKey_2` -> `Mnemonic_2`.
With the above data structure, it's difficult to find which backup comes
first when the client imports the SecretData in the frontend.

So, in this PR, the return values of `fetchAllSecretData` are updated to
the Array of SecretMetadata objects, for e.g.
```ts
[
  { data: <Mnemonic_1>, type: Mnemonic, timestamp: <Timestamp-1> },
  { data: <Mnemonic_2>, type: Mnemonic, timestamp: <Timestamp-2> },
  { data: <PrivateKey_2>, type: PrivateKey, timestamp: <Timestamp-3> },
  { data: <Mnemonic_3>, type: Mnemonic, timestamp: <Timestamp-4> },
  ...
]
```
This data structure will help the clients to know in which order the
secret data were imported as well as includes the type, so that the
client can process the relevant imports.

Beside this main change, additional validations were added in the
`fetchAllSecretData` method, which are ~
- Throws error when the client receives the empty array of `SecretData`
(which is impossible in working scenario).
- Throws error when the first `SecretData` is not `Mnemonics` type (the
first secretData backup must always be the Mnemonics).
Moreover, the PR includes some internal refactoring (mostly on the
`fetchAllSecretData` method), because with the new validations added and
some logic changes, the method becomes a bit harder to read.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: matthiasgeihs <62935430+matthiasgeihs@users.noreply.github.com>