fix: use scopes to retrieve chain ID (#6072)
## Explanation

We currently use the address to identify what network the account
belongs to, yet that method is incorrect as:
- accounts span multiple networks 
- addresses can be the same on different networks.

The current implementation leads to a bug where the network will always
be Mainnet for Solana and Bitcoin, preventing testnets to work. It would
also fail to work on EVM networks if this function was used.

The solution, using scopes, fixes the problem for Bitcoin and introduces
no regression for Solana.

## References

* Fixes https://consensyssoftware.atlassian.net/browse/NWNT-397