Improve create-update-issues workflow
We have a workflow to automatically create tickets in both Extension and
Mobile whenever a new major version of a package is published. However,
there are a few of issues with this workflow:

- The command to create the issue can fail if a team label doesn't exist
  in either Extension or Mobile. This seems to prevent other issues from
  being created.
- It's actually difficult to tell which issues get successfully created
  and which don't, because there isn't much output from this workflow.
- The description in the issue is very terse. It says "Please update
  this package", assigns it to a team, and doesn't explain why that
  should care or what they should do next.
- The script in the workflow is difficult to maintain because it's
  inline.
- It's difficult to test the script without forking the repo.

This commit extracts the script to a separate file and addresses the
above issues by adding better failure handling and logging, expanding
the issue description, and adding a dry-run mode.

To test the changes here, run:

```
./scripts/create-update-issues.sh --ref v460.0.0

./scripts/create-update-issues.sh --ref v459.0.0
```