Improve create-update-issues workflow (#6090)
We have a workflow to automatically create tickets in both Extension and
Mobile whenever a new major version of a package is published. However,
there are some issues with this workflow:

- Some of the team labels referenced in `teams.json` will be or have
already changed in the Extension and Mobile repos.
- If if a team label does _not_ exist in either Extension or Mobile, the
command to create the issue can fail. This seems to prevent other issues
from being created.
- If the workflow fails midway through and is run again, then it can
create duplicate issues.
- It's actually difficult to tell which issues get successfully created
and which don't, because there isn't much output from this workflow.
- The description in the issue is very terse. It says "Please update
this package", assigns it to a team, and doesn't explain why they should
care or what they should do next.
- The script in the workflow is difficult to maintain because it's
inline.
- It's difficult to test the script without forking the repo.

This commit extracts the script to a separate file and addresses the
above issues by adding better failure handling and logging, expanding
the issue description, adding a dry-run mode, and ensuring that
duplicate issues do not get created.

---------

Co-authored-by: Mark Stacey <markjstacey@gmail.com>