feat: Stop persisting `permissionActivityLog` state
The `permissionActivityLog` property of the `PermissionLogController`
state tracks all calls to restricted methods. This was intended as a
diagnostic aid, but in practice it is rarely used. This log is causing
performance issues related to disk writes because it can be triggered
while the wallet is idle (by snaps calling restricted methods).

To reduce the performance burden of this property, it is no longer
persisted. This list will still be available in-memory for diagnosing
problems, but will not be preserved across restarts.

Relates to https://github.com/MetaMask/metamask-extension/issues/33879