feat: Stop persisting `permissionActivityLog` state (#6156)
## Explanation

The `permissionActivityLog` property of the `PermissionLogController`
state tracks all calls to restricted methods. This was intended as a
diagnostic aid, but in practice it is rarely used. This log is causing
performance issues related to disk writes because it can be triggered
while the wallet is idle (by snaps calling restricted methods).

To reduce the performance burden of this property, it is no longer
persisted. This list will still be available in-memory for diagnosing
problems, but will not be preserved across restarts.

## References

Relates to https://github.com/MetaMask/metamask-extension/issues/33879

Here is the draft PR of this update for `metamask-extension` that
addresses the breaking changes:
https://github.com/MetaMask/metamask-extension/pull/34465
This controller is not used by mobile, so there is no draft PR for
mobile.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes