fix: Update access token validation and retrieval from vault and state  (#6155)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

This PR includes fix and update to get the `accessToken` value from the
controller.
`accessToken` data is saved in both vault and state (as non-persisted
value) from the `OAuth Authentication` and previously, we only read and
validate it from the state when updating the SeedlessOnboarding vault
data.

However, there's an edge case in the client (especially in the
extension) where `acessToken` is persisted to vault but removed from the
state. Please refer to this extension bug
(https://github.com/MetaMask/metamask-extension/issues/34437) as an
example.
In that case, `accessToken` value is persisted to vault and set
(temporarily in mem) to state after the rehydration (unlock), however
when the user refresh the page before the onboarding is finished, the
`accessToken` value is removed from the state(memory) and when user
tries to rehydrate, the controller throw the `InvalidAccessToken` error
as it's not available in the in-memory state anymore.

This PR fixes the issue by checking the access-token in both in-memory
state and in the encrypted vault.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

Fixes: https://github.com/MetaMask/metamask-extension/issues/34437

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes