Release 488.0.0 (#6249)
## Explanation

This PR releases version 73.0.2 of the `@metamask/assets-controllers`
package. This is a patch release that includes a fix to use a narrow
selector when listening to `CurrencyRateController:stateChange`, as well
as a fix to an issue where the `MultichainAssetsRatesController` was
attempting to fetch asset conversions for accounts without assets,
causing the snap to crash.

## References

- #6217
- #6207

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes