refactor(account-tree-controller): remove `wallet.metadata.entropy.index` (#6258)
## Explanation

Removing the `wallet.metadata.entropy.index`. This is mainly used by us
for the naming, and the "order" is the responsibility of the
`KeyringController` mainly, not the `AccountTreeController`.

## References

N/A

## Changelog

N/A

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes