refactor(account-tree-controller)!: remove `AccountTree{Wallet,Group}` in-memory instances (#6265)
## Explanation

This implementation is not good enough to start using it.

We will re-enable this once we have a better solution for those. For the
moment, we will add actions to the controller to be able to fetch
accounts from the selected account group and that should cover most
use-cases for the moment.

They will be replaced by:
- https://github.com/MetaMask/core/pull/6266

## References

N/A

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes