chore: Fix internal type mismatch (#6264)
## Explanation

Fix a type mismatch between the `subscribe` method and the override for
it that accepts a selector. This didn't cause a type error and it
doesn't impact users of this package, but leaving this unresolved left
us at risk because we're type checking with the wrong parameter type.

## References

N/A

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes