Add `error` property to "degraded" events (#6188)
We would like to start tracking the HTTP status code in metrics when the
`NetworkController:rpcEndpointDegraded` event is published because the
maximum number of retries for a request has been exceeded. The HTTP
status is stored in the error that Cockatiel captures, so we just need
to include it in the event payload.