chore(network-controller): Remove lookupNetworkByClientId; refactor lookupNetwork
A while back, `lookupNetworkByClientId` was added a while back,
presumably as a multichain-compatible replacement of `lookupNetwork`.
However, `lookupNetwork` already takes an (optional) network client ID;
in fact, it calls `lookupNetworkByClientId` internally. This duplication
is confusing, and it would make for a simpler API and make
NetworkController more maintainable if we removed
`lookupNetworkByClientId` entirely.

This commit does that and refactors `lookupNetwork` to be more
understandable. It also ensures that `lookupNetwork` is well-tested both
when given a network client ID and not given one.