feat: add alignment methods to group, wallet and service classes (#6326)
## Explanation

**What is the current state of things and why does it need to change?**
An alignment mechanism didn't exist for multichain account groups. It
needs to be added because we can have multiple scenarios where
misalignment occurs:
1. State 1 -> State 2 - We will have to align existing account groups
when they are created.
2. New protocols added - Existing groups will have to have an account
associated with the new provider (chain).
3. Basic functionality toggle - With basic functionality off, we can
have a scenario where only an EVM account is created when a group is
created.
4. Discovery - When importing an SRP, the discovery process might not
necessarily yield a discovered account for every provider and we have to
align the group in those scenarios.

**What is the solution your changes offer and how does it work?** It
adds an alignment mechanism at the group level and it runs through the
list of providers and checks to see if any accounts exist for that
particular group, if there aren't any, then it calls
`provider.createAccounts` with the associated entropy id and group
index. Alignment methods were added at the wallet level where a wallet
would call align on all of its existing groups and similarly at the
service level where it can call align on all of its wallets. Singular
methods were also added at the wallet and service level where a
particular group associated with a wallet can be aligned and a
particular wallet can be aligned by the service.

Note: I also refactored the `withCreateAccount` in `SnapAccountProvider`
to be a bit more readable and easier to understand.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes